# docs/v1/swagger.yaml
definitions:
...
  dto.InvoiceRequest:
    properties:
      due_date:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  dto.InvoiceResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      due_date:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.TransactionCategoryResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.TransactionInvoiceResponse:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
...
  dto.TransactionResponse:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/dto.TransactionCategoryResponse'
      created_at:
        type: string
      id:
        type: string
      invoice:
        $ref: '#/definitions/dto.TransactionInvoiceResponse'
      record_date:
        type: string
      record_type:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  title: API Frog-Go
  version: "1.0"
paths:
...
  /api/v1/invoices:
    get:
      consumes:
      - application/json
      description: Lista todas as faturas aplicando filtros e paginação
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Limite por página
        in: query
        name: limit
        type: integer
      - description: 'Campo de ordenação (ex: due_date)'
        in: query
        name: order_by
        type: string
      - description: Ordem (asc, desc)
        in: query
        name: order
        type: string
      - description: Filtro por título
        in: query
        name: title
        type: string
      - description: Filtro por status
        in: query
        name: status
        type: string
      - description: Valor mínimo
        in: query
        name: min_amount
        type: number
      - description: Valor máximo
        in: query
        name: max_amount
        type: number
      - description: Data de vencimento inicial
        in: query
        name: due_date_start
        type: string
      - description: Data de vencimento final
        in: query
        name: due_date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.InvoiceResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista faturas com filtros e paginação
      tags:
      - Faturas
    post:
      consumes:
      - application/json
      description: Cria uma nova fatura com os dados fornecidos no corpo da requisição
      parameters:
      - description: Dados da fatura
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Cria uma nova fatura
      tags:
      - Faturas
  /api/v1/invoices/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui uma fatura com base no ID fornecido
      parameters:
      - description: ID da fatura
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Sem conteúdo
      security:
      - BearerAuth: []
      summary: Remove uma fatura
      tags:
      - Faturas
    get:
      consumes:
      - application/json
      description: Retorna os dados de uma fatura com base no ID fornecido
      parameters:
      - description: ID da fatura
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Busca uma fatura por ID
      tags:
      - Faturas
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma fatura com base no ID fornecido e nos
        dados enviados no corpo da requisição
      parameters:
      - description: ID da fatura
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados da fatura
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
      security:
      - BearerAuth: []
      summary: Atualiza uma fatura existente
      tags:
      - Faturas
  /api/v1/invoices/{id}/debts:
    get:
      consumes:
      - application/json
      description: Lista todos os débitos (transações) vinculados a uma fatura específica,
        com filtros e paginação
      parameters:
      - description: ID da fatura
        in: path
        name: id
        required: true
        type: string
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Limite por página
        in: query
        name: limit
        type: integer
      - description: 'Campo de ordenação (ex: record_date)'
        in: query
        name: order_by
        type: string
      - description: Ordem (asc, desc)
        in: query
        name: order
        type: string
      - description: Filtro por título da transação
        in: query
        name: title
        type: string
      - description: Filtro por status da transação
        in: query
        name: status
        type: string
      - description: Tipo da transação (income, expense)
        in: query
        name: record_type
        type: string
      - description: ID da categoria
        in: query
        name: category_id
        type: string
      - description: Data inicial de registro
        in: query
        name: record_date_start
        type: string
      - description: Data final de registro
        in: query
        name: record_date_end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TransactionResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Lista débitos associados a uma fatura
      tags:
      - Faturas
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT no formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"